requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor==6.7.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
